+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.3] - 2024-10-14
+## [0.0.5] - 2024-10-24
+
+- Printify order integration (mvp, bare bones)
+
+## [0.0.4] - 2024-10-23
+
+- Stripe checkout functionality (w/ async webhook)
+
+## [0.0.3] - 2024-10-19
 
 - Client-side cart item storage
 - Catalog filters/pagination